+++ stm32/ros_usbnode/platformio.ini
@@ -14,6 +14,11 @@ default_envs = Yardforce500
 [env]
 platform = ststm32
 framework = stm32cube
+build_src_filter =
+	+<*>
+	-<.git/>
+	-<.svn/>
+	-<proxy_inc/**/*.c>
 
 [env:Yardforce500]
 board = genericSTM32F103VC
@@ -23,8 +28,7 @@ platform_packages =
 	platformio/tool-openocd@^2.1100.211028
 	platformio/tool-dfuutil@^1.11.0
 build_flags = -DBOARD_YARDFORCE500_VARIANT_ORIG=1 -Wl,--undefined,_printf_float  -O2 -Isrc/ros/ros_lib -Isrc/ros/ros_custom
-build_src_filter = 
-extra_scripts = 
+extra_scripts =
 	pre:patch_usb.py
 	pre:add_swo_viewer.py
 debug_tool = stlink
@@ -39,7 +43,6 @@ platform_packages =
 	platformio/tool-openocd@^2.1100.211028
 	platformio/tool-dfuutil@^1.11.0
 build_flags = -DBOARD_YARDFORCE500_VARIANT_B=1 -Wl,--undefined,_printf_float  -O2 -Isrc/ros/ros_lib -Isrc/ros/ros_custom
-build_src_filter =
 extra_scripts =
 	pre:patch_usb.py
 	pre:add_swo_viewer.py
