+++ stm32/ros_usbnode/src/charger.c
@@ -139,7 +139,11 @@ uint8_t  chargecontrol_is_charging  = 0;
   GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
   GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
   HAL_GPIO_Init(CHARGE_GPIO_PORT, &GPIO_InitStruct);
+
+#if BOARD_YARDFORCE500_VARIANT_ORIG
+  // TODO: Is something equivalent needed for the STM32f4?
   __HAL_AFIO_REMAP_TIM1_ENABLE();        // to use PE8/9 it is a full remap
+#endif
 
 
     // Charge CH1/CH1N PWM Timer
