+++ stm32/ros_usbnode/src/proxy_inc/stm32f4/stm32f4xx_it.c
@@ -1,0 +1,399 @@
+/* USER CODE BEGIN Header */
+/**
+  ******************************************************************************
+  * @file    stm32f4xx_it.c
+  * @brief   Interrupt Service Routines.
+  ******************************************************************************
+  * @attention
+  *
+  * Copyright (c) 2022 STMicroelectronics.
+  * All rights reserved.
+  *
+  * This software is licensed under terms that can be found in the LICENSE file
+  * in the root directory of this software component.
+  * If no LICENSE file comes with this software, it is provided AS-IS.
+  *
+  ******************************************************************************
+  */
+/* USER CODE END Header */
+
+/* Includes ------------------------------------------------------------------*/
+#include "board.h"
+#include "main.h"
+#include "panel.h"
+#include "stm32f1xx_it.h"
+/* Private includes ----------------------------------------------------------*/
+/* USER CODE BEGIN Includes */
+/* USER CODE END Includes */
+
+/* Private typedef -----------------------------------------------------------*/
+/* USER CODE BEGIN TD */
+
+/* USER CODE END TD */
+
+/* Private define ------------------------------------------------------------*/
+/* USER CODE BEGIN PD */
+
+/* USER CODE END PD */
+
+/* Private macro -------------------------------------------------------------*/
+/* USER CODE BEGIN PM */
+
+/* USER CODE END PM */
+
+/* Private variables ---------------------------------------------------------*/
+/* USER CODE BEGIN PV */
+
+uint16_t cnt_uart4_overrun = 0;
+uint16_t cnt_usart2_overrun = 0;
+
+/* USER CODE END PV */
+
+/* Private function prototypes -----------------------------------------------*/
+/* USER CODE BEGIN PFP */
+
+/* USER CODE END PFP */
+
+/* Private user code ---------------------------------------------------------*/
+/* USER CODE BEGIN 0 */
+
+/* USER CODE END 0 */
+
+/* External variables --------------------------------------------------------*/
+extern UART_HandleTypeDef MASTER_USART_Handler;
+extern UART_HandleTypeDef DRIVEMOTORS_USART_Handler;
+extern UART_HandleTypeDef BLADEMOTOR_USART_Handler;
+extern PCD_HandleTypeDef hpcd_USB_FS;
+
+
+extern DMA_HandleTypeDef hdma_uart1_tx;
+extern DMA_HandleTypeDef hdma_uart1_rx;
+extern DMA_HandleTypeDef hdma_usart2_rx;
+extern DMA_HandleTypeDef hdma_usart2_tx;
+extern DMA_HandleTypeDef hdma_uart_blade_tx;
+extern DMA_HandleTypeDef hdma_uart_blade_rx;
+extern DMA_HandleTypeDef hdma_uart4_tx;
+extern DMA_HandleTypeDef hdma_uart4_rx;
+extern DMA_HandleTypeDef hdma_adc;
+
+extern ADC_HandleTypeDef ADC_Charging_Handle;
+
+/* USER CODE BEGIN EV */
+
+/* USER CODE END EV */
+
+/******************************************************************************/
+/*           Cortex-M3 Processor Interruption and Exception Handlers          */
+/******************************************************************************/
+/**
+  * @brief This function handles Non maskable interrupt.
+  */
+void NMI_Handler(void)
+{
+	/* USER CODE BEGIN NonMaskableInt_IRQn 0 */
+
+	/* USER CODE END NonMaskableInt_IRQn 0 */
+	/* USER CODE BEGIN NonMaskableInt_IRQn 1 */
+	while (1)
+	{
+	}
+	/* USER CODE END NonMaskableInt_IRQn 1 */
+}
+
+/**
+  * @brief This function handles Hard fault interrupt.
+  */
+void HardFault_Handler(void)
+{
+	/* USER CODE BEGIN HardFault_IRQn 0 */
+
+	/* USER CODE END HardFault_IRQn 0 */
+	while (1)
+	{
+		/* USER CODE BEGIN W1_HardFault_IRQn 0 */
+		/* USER CODE END W1_HardFault_IRQn 0 */
+	}
+}
+
+/**
+  * @brief This function handles Memory management fault.
+  */
+void MemManage_Handler(void)
+{
+	/* USER CODE BEGIN MemoryManagement_IRQn 0 */
+
+	/* USER CODE END MemoryManagement_IRQn 0 */
+	while (1)
+	{
+		/* USER CODE BEGIN W1_MemoryManagement_IRQn 0 */
+		/* USER CODE END W1_MemoryManagement_IRQn 0 */
+	}
+}
+
+/**
+  * @brief This function handles Prefetch fault, memory access fault.
+  */
+void BusFault_Handler(void)
+{
+	/* USER CODE BEGIN BusFault_IRQn 0 */
+
+	/* USER CODE END BusFault_IRQn 0 */
+	while (1)
+	{
+		/* USER CODE BEGIN W1_BusFault_IRQn 0 */
+		/* USER CODE END W1_BusFault_IRQn 0 */
+	}
+}
+
+/**
+  * @brief This function handles Undefined instruction or illegal state.
+  */
+void UsageFault_Handler(void)
+{
+	/* USER CODE BEGIN UsageFault_IRQn 0 */
+
+	/* USER CODE END UsageFault_IRQn 0 */
+	while (1)
+	{
+		/* USER CODE BEGIN W1_UsageFault_IRQn 0 */
+		/* USER CODE END W1_UsageFault_IRQn 0 */
+	}
+}
+
+/**
+  * @brief This function handles System service call via SWI instruction.
+  */
+void SVC_Handler(void)
+{
+	/* USER CODE BEGIN SVCall_IRQn 0 */
+
+	/* USER CODE END SVCall_IRQn 0 */
+	/* USER CODE BEGIN SVCall_IRQn 1 */
+
+	/* USER CODE END SVCall_IRQn 1 */
+}
+
+/**
+  * @brief This function handles Debug monitor.
+  */
+void DebugMon_Handler(void)
+{
+	/* USER CODE BEGIN DebugMonitor_IRQn 0 */
+
+	/* USER CODE END DebugMonitor_IRQn 0 */
+	/* USER CODE BEGIN DebugMonitor_IRQn 1 */
+
+	/* USER CODE END DebugMonitor_IRQn 1 */
+}
+
+/**
+  * @brief This function handles Pendable request for system service.
+  */
+void PendSV_Handler(void)
+{
+	/* USER CODE BEGIN PendSV_IRQn 0 */
+
+	/* USER CODE END PendSV_IRQn 0 */
+	/* USER CODE BEGIN PendSV_IRQn 1 */
+
+	/* USER CODE END PendSV_IRQn 1 */
+}
+
+/**
+  * @brief This function handles System tick timer.
+  */
+void SysTick_Handler(void)
+{
+	/* USER CODE BEGIN SysTick_IRQn 0 */
+
+	/* USER CODE END SysTick_IRQn 0 */
+	HAL_IncTick();
+	/* USER CODE BEGIN SysTick_IRQn 1 */
+
+	/* USER CODE END SysTick_IRQn 1 */
+}
+
+/******************************************************************************/
+/* STM32F1xx Peripheral Interrupt Handlers                                    */
+/* Add here the Interrupt Handlers for the used peripherals.                  */
+/* For the available peripheral interrupt handler names,                      */
+/* please refer to the startup file (startup_stm32f1xx.s).                    */
+/******************************************************************************/
+
+/**
+  * @brief This function handles ADC 1 & 2 global interrupt.
+  */
+void ADC1_2_IRQHandler(void)
+{
+	HAL_ADC_IRQHandler(&ADC_Charging_Handle);
+}
+
+
+/**
+  * @brief This function handles USART1 global interrupt.
+  */
+void USART1_IRQHandler(void)
+{
+	/* USER CODE BEGIN USART1_IRQn 0 */
+
+	/* USER CODE END USART1_IRQn 0 */
+	HAL_UART_IRQHandler(&PANEL_USART_Handler);
+
+	/* USER CODE BEGIN USART1_IRQn 1 */
+
+	/* USER CODE END USART1_IRQn 1 */
+}
+
+/**
+  * @brief This function handles UART global interrupt. (DRIVE MOTOR)
+  */
+void USART2_IRQHandler(void)
+{
+
+	uint32_t status = USART2->SR;
+	if (status & USART_SR_ORE){ // overrun error
+		cnt_usart2_overrun++;
+	}
+
+	HAL_UART_IRQHandler(&DRIVEMOTORS_USART_Handler);
+}
+
+/**
+ * @brief This function handles UART global interrupt. (BLADE MOTOR)
+ */
+void USART3_IRQHandler(void)
+{
+	HAL_UART_IRQHandler(&BLADEMOTOR_USART_Handler);
+}
+
+
+/**
+  * @brief This function handles UART4 global interrupt.
+  */
+void UART4_IRQHandler(void)
+{
+	HAL_UART_IRQHandler(&MASTER_USART_Handler);
+}
+
+/**
+  * @brief This function handles DMA1 channel1 global interrupt.
+  */
+void DMA1_Channel1_IRQHandler(void)
+{
+	/* USER CODE BEGIN DMA1_Channel1_IRQn 0 */
+
+	/* USER CODE END DMA1_Channel1_IRQn 0 */
+	HAL_DMA_IRQHandler(&hdma_adc);
+	/* USER CODE BEGIN DMA1_Channel1_IRQn 1 */
+
+	/* USER CODE END DMA1_Channel1_IRQn 1 */
+}
+
+/**
+  * @brief This function handles DMA1 channel2 global interrupt.
+  */
+void DMA1_Channel2_IRQHandler(void)
+{
+	HAL_DMA_IRQHandler(&hdma_uart_blade_tx);
+}
+
+/**
+  * @brief This function handles DMA1 channel3 global interrupt.
+  */
+void DMA1_Channel3_IRQHandler(void)
+{
+	HAL_DMA_IRQHandler(&hdma_uart_blade_rx);
+}
+
+/**
+  * @brief This function handles DMA1 channel4 global interrupt.
+  */
+void DMA1_Channel4_IRQHandler(void)
+{
+	HAL_DMA_IRQHandler(&hdma_uart1_tx);
+}
+
+/**
+  * @brief This function handles DMA1 channel5 global interrupt.
+  */
+void DMA1_Channel5_IRQHandler(void)
+{
+	HAL_DMA_IRQHandler(&hdma_uart1_rx);
+}
+
+
+/**
+  * @brief This function handles DMA1 channel6 global interrupt. (DRIVE MOTOR UART)
+  */
+void DMA1_Channel6_IRQHandler(void)
+{
+	/* USER CODE BEGIN DMA1_Channel6_IRQn 0 */
+
+	/* USER CODE END DMA1_Channel6_IRQn 0 */
+	HAL_DMA_IRQHandler(&hdma_usart2_rx);
+	/* USER CODE BEGIN DMA1_Channel6_IRQn 1 */
+
+	/* USER CODE END DMA1_Channel6_IRQn 1 */
+}
+
+/**
+  * @brief This function handles DMA1 channel7 global interrupt. (DRIVE MOTOR UART)
+  */
+void DMA1_Channel7_IRQHandler(void)
+{
+	/* USER CODE BEGIN DMA1_Channel7_IRQn 0 */
+
+	/* USER CODE END DMA1_Channel7_IRQn 0 */
+	HAL_DMA_IRQHandler(&hdma_usart2_tx);
+	/* USER CODE BEGIN DMA1_Channel7_IRQn 1 */
+
+	/* USER CODE END DMA1_Channel7_IRQn 1 */
+}
+
+/**
+  * @brief This function handles DMA2 channel3 global interrupts.
+  */
+void DMA2_Channel3_IRQHandler(void)
+{
+	/* USER CODE BEGIN DMA2_Channel3_IRQn 0 */
+
+	/* USER CODE END DMA2_Channel3_IRQn 0 */
+	HAL_DMA_IRQHandler(&hdma_uart4_rx);
+	/* USER CODE BEGIN DMA2_Channel3_IRQn 1 */
+
+	/* USER CODE END DMA2_Channel3_IRQn 1 */
+}
+
+/**
+  * @brief This function handles DMA2 channel4 and channel5 global interrupts.
+  */
+void DMA2_Channel4_5_IRQHandler(void)
+{
+	/* USER CODE BEGIN DMA2_Channel4_5_IRQn 0 */
+
+	/* USER CODE END DMA2_Channel4_5_IRQn 0 */
+	HAL_DMA_IRQHandler(&hdma_uart4_tx);
+	/* USER CODE BEGIN DMA2_Channel4_5_IRQn 1 */
+
+	/* USER CODE END DMA2_Channel4_5_IRQn 1 */
+}
+
+
+
+/**
+  * @brief This function handles USB low priority or CAN RX0 interrupts.
+  */
+void USB_LP_CAN1_RX0_IRQHandler(void)
+{
+	/* USER CODE BEGIN USB_LP_CAN1_RX0_IRQn 0 */
+
+	/* USER CODE END USB_LP_CAN1_RX0_IRQn 0 */
+	HAL_PCD_IRQHandler(&hpcd_USB_FS);
+	/* USER CODE BEGIN USB_LP_CAN1_RX0_IRQn 1 */
+
+	/* USER CODE END USB_LP_CAN1_RX0_IRQn 1 */
+}
+
+/* USER CODE BEGIN 1 */
+
+/* USER CODE END 1 */
