+++ stm32/ros_usbnode/src/startup_stm32f.s
@@ -1,0 +1,5 @@
+#if BOARD_YARDFORCE500_VARIANT_ORIG
+#include "proxy_inc/stm32f1/startup_stm32f1xx.s"
+#elif BOARD_YARDFORCE500_VARIANT_B
+#include "proxy_inc/stm32f4/startup_stm32f4xx.s"
+#endif
