+++ stm32/ros_usbnode/src/stm32f_it.c
@@ -1,0 +1,5 @@
+#if BOARD_YARDFORCE500_VARIANT_ORIG
+#include "proxy_inc/stm32f1/stm32f1xx_it.c"
+#elif BOARD_YARDFORCE500_VARIANT_B
+#include "proxy_inc/stm32f4/stm32f4xx_it.c"
+#endif
